refactor: final 처리 추가
