+++ src/main/java/org/sopt/carena/member/adapter/in/web/dto/response/MyPageResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.carena.member.adapter.in.web.dto;
+package org.sopt.carena.member.adapter.in.web.dto.response;
 
 import org.sopt.carena.member.application.dto.view.MyPageInfoView;
 
@@ -8,7 +8,7 @@ public record MyPageResponse(
         String name,
         LocalDate birthdate
 ) {
-    public static MyPageResponse from(MyPageInfoView view) {
+    public static MyPageResponse from(final MyPageInfoView view) {
         return new MyPageResponse(
                 view.name(),
                 view.birthdate()
