+++ src/main/java/org/sopt/carena/member/application/dto/view/MyPageInfoView.java
@@ -6,7 +6,7 @@ public record MyPageInfoView(
         String name,
         LocalDate birthdate
 ) {
-    public static MyPageInfoView of(String name, LocalDate birthdate) {
+    public static MyPageInfoView of(final String name, final LocalDate birthdate) {
         return new MyPageInfoView(
                 name, birthdate);
     }
